chore: modified layout
