+++ frontend/src/components/EconomicSection.tsx
@@ -12,7 +12,7 @@ export const EconomicSection = ({ articles, loading }: EconomicSectionProps) => 
     if (!loading && (!articles || articles.length === 0)) return null;
 
     return (
-        <div className="container-main py-8 border-t border-gray-100 mt-4">
+        <div className="container-main py-8 mt-4">
             <div className="flex mb-6">
                 <h2 className="bg-[#cc0000] text-white px-4 py-1.5 text-[13px] font-bold uppercase tracking-wide">
                     DOANH NGHIỆP - SẢN PHẨM - DỊCH VỤ
