+++ frontend/src/index.css
@@ -22,7 +22,6 @@
 body {
     font-family: var(--font-body);
     color: var(--color-text-primary);
-    background-color: var(--color-bg-light);
     line-height: 1.6;
 }
